IVFPSNR: Disable SSIM output if not result not used.
Don't calculate the SSIM if the result isn't used by the test.

Current SSIM implementation is s-l-o-w and was dominating test
execution time, but the result was never used in a test. In fact,
it seems that all tests that purport to use SSIM may be broken,
since no PNSR function seems to assign to the SsimOut parameter
that is passed in. That can be reexamined in a later commit.

Change-Id: Icd29e564c5ff31fcb50226afcf35bded2bb83b2c