Updated how mode set for compressions and WinLinMacMatch
-Compressions no longer require mode 3 to be independently run
prior to second pass so the method used to set mode could be
simplified in many cases. Some legacy method still exists for
new vs old tests to ensure reverse compatibility.

-Updated Window Matches Linux to WinLinMacMatches now uses filer
and version numbers to store and check compression and decompression
equivalence across platforms.  Test will only return pass or fail if
compressions and decompression exist on filer for the version number
being run for the two platforms it is not running on.  The test will
return Indeterminate if test files only are made.

Change-Id: I2dc20df24fb24f888a0b4e276d6ce65b532152de